Revert "refactor(*): move commands into directory (#9)"
This reverts commit 65ec13c58f6b9bd0e3ba9ed488074052c65a8838.
